pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]

name = "jnp3"
description = "一个 Python 语言的辅助工具集"
authors = [
    {name = "Julian Freeman", email = "freemanjulian372@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}

dependencies = []
requires-python = ">=3.10"

[project.optional-dependencies]
gui = [
    "PySide6-Essentials==6.5.2"
]

[tool.setuptools.dynamic]
version = {attr = "jnp3.__version__"}
